docker/requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.10
bunch==1.0.1
click==3.3
docker-crane==0.1.0
dockerpty==0.2.3
docopt==0.6.1
fig==0.5.2
pyaml==14.05.7
requests==2.2.1
texttable==0.8.1
websocket-client==0.11.0
wsgiref==0.1.2
